+++ auth/CHANGES.rst
@@ -1,5 +1,6 @@
 0.0
 ---
 
-- Added tests for public, protected, privileged pages
-- Added data models
+- Made tests for public, protected, privileged pages
+- Made data models
+- Made public, protected, privileged pages